updated breathing
