+++ addons/acebreathing/$PBOPREFIX$
@@ -1,1 +1,1 @@
-kat_acebreathing
+x\kat_aceAdvMedical\addons\acebreathing
