--- addons/acebreathing/BIS_AddonInfo.hpp
@@ -1,4 +1,0 @@
-class BIS_AddonInfo
-{
-    author="Katalam";
-};
