+++ addons/acebreathing/functions/fnc_init.sqf
@@ -1,3 +1,4 @@
+#include "script_component.hpp"
 /*
  * Author: Katalam
  * Initializes unit variables.
@@ -9,13 +10,13 @@
  * None
  *
  * Example:
- * [bob] call kat_aceBreathing_fnc_init
+ * [bob] call kat_acebreathing_fnc_init
  *
  * Public: No
  */
 
 params ["_unit"];
 
-_unit setVariable ["kat_aceBreathing_pulseoximeter", false, true];
-_unit setVariable ["kat_aceBreathing_o2", false, true];
+_unit setVariable [QGVAR(pulseoximeter), false, true];
+_unit setVariable [QGVAR(o2), false, true];
 // ace_medical_airwayStatus - oxygen saturation
