+++ addons/acebreathing/functions/fnc_treatmentAdvanced_pulseoximeter.sqf
@@ -1,3 +1,4 @@
+#include "script_component.hpp"
 /*
  * Author: Katalam
  * docks a pulseoximeter on the patient
@@ -11,7 +12,7 @@
  * None
  *
  * Example:
- * call kat_aceBreathing_fnc_treatmentAdvanced_pulseoximeter;
+ * call kat_acebreathing_fnc_treatmentAdvanced_pulseoximeter;
  *
  * Public: No
  */
@@ -19,7 +20,7 @@
 params ["_player", "_target"];
 
 if (local _target) then {
-  ["treatmentPulseoximeter", [_player, _target]] call CBA_fnc_localEvent;
+    ["treatmentPulseoximeter", [_player, _target]] call CBA_fnc_localEvent;
 } else {
-  ["treatmentPulseoximeter", [_player, _target], _target] call CBA_fnc_targetEvent;
+    ["treatmentPulseoximeter", [_player, _target], _target] call CBA_fnc_targetEvent;
 };
