+++ addons/acebreathing/functions/script_component.hpp
@@ -1,0 +1,1 @@
+#include "\x\kat_aceAdvMedical\addons\acebreathing\script_component.hpp"
