+++ addons/acebreathing/models/pulseox/pulseox.rvmat
@@ -12,7 +12,7 @@ PixelShaderID="Super";
 VertexShaderID="Super";
 class Stage1
 {
-    texture="kat_acebreathing\pulseox\pulseox_nohq.paa";
+    texture="x\kat_aceAdvMedical\addons\acebreathing\models\pulseox\pulseox_nohq.paa";
     uvSource="tex";
     class uvTransform
     {
