+++ addons/acebreathing/script_component.hpp
@@ -1,0 +1,17 @@
+#define COMPONENT acebreathing
+#define COMPONENT_BEAUTIFIED KAT - ACE Breathing
+#include "\x\kat_aceAdvMedical\main\script_mod.hpp"
+
+// #define DEBUG_MODE_FULL
+// #define DISABLE_COMPILE_CACHE
+// #define ENABLE_PERFORMANCE_COUNTERS
+
+#ifdef DEBUG_ENABLED_BREATHING
+    #define DEBUG_MODE_FULL
+#endif
+
+#ifdef DEBUG_SETTINGS_BREATHING
+    #define DEBUG_SETTINGS DEBUG_SETTINGS_BREATHING
+#endif
+
+#include "\x\kat_aceAdvMedical\main\script_macros.hpp"
